DentalBridge/backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
google-generativeai
pdfplumber
sqlalchemy
pytesseract
pdf2image
python-dotenv
